LayoutTests/inspector/debugger-pause-on-debugger-statement.html

<html>
<head>
<script src="../http/tests/inspector/inspector-test.js"></script>
<script>

function testFunction() {
    debugger;
}

function doit()
{
    function callback(result)
    {
        if (result instanceof Array) {
            for (var i = 0; i < result.length; i++)
                output(result[i]);
        } else
            output(result);
        notifyDone();
    }
    evaluateInWebInspector("frontend_testPauseOnDebuggerStatement", callback);
}

function frontend_testPauseOnDebuggerStatement(testController)
{
    testController.waitUntilDone();

    WebInspector.showScriptsPanel();
    frontend_ensureDebuggerEnabled(debuggerWasEnabledCallback);

    function debuggerWasEnabledCallback()
    {
        testController.results.push("Debugger was enabled.");
        frontend_evaluateInConsole("setTimeout(testFunction, 0)", function(result)
        {
            testController.results.push("Set timer for test function.");
        });
    }

    frontend_addSniffer(WebInspector, "pausedScript", function(callFrames)
    {
        testController.results.push("Script execution paused.");
        if (WebInspector.panels.scripts.paused)
            WebInspector.panels.scripts._togglePause();
        testController.results.push("Resumed script execution.");
        testController.notifyDone();
    });
}

function frontend_ensureDebuggerEnabled(callback)
{
    if (WebInspector.panels.scripts._debuggerEnabled) {
        callback();
        return;
    }

    frontend_addSniffer(WebInspector, "debuggerWasEnabled", callback);
    WebInspector.panels.scripts._toggleDebugging(false);
}

</script>
</head>

<body onload="onload()">
<p>
Tests that debugger will stop on "debugger" statement.
</p>

</body>
</html>
