LayoutTests/inspector/debugger-pause-on-debugger-statement-expected.txt

Tests that debugger will stop on "debugger" statement.

Debugger was enabled.
Set timer for test function.
Script execution paused.
Resumed script execution.
